third_party/WebKit/LayoutTests/http/tests/security/video-poster-cross-origin-crash-expected.txt

CONSOLE ERROR: Cross-origin image load denied by Cross-Origin Resource Sharing policy.
CONSOLE ERROR: Cross-origin image load denied by Cross-Origin Resource Sharing policy.
>>>
